+++ shared/src/index.ts
@@ -4,6 +4,7 @@ export * from "./valuation";
 export * from "./player";
 export * from "./goalkeeperGrid";
 export * from "./probableLineup";
+export * from "./setPieceTaker";
 export * from "./manager";
 export * from "./purchase";
 export * from "./maxBid";